+++ src/vmaf.rs
@@ -1,5 +1,5 @@
 //! vmaf logic
-use crate::process::{cmd_err, exit_ok_stderr, Chunks, CommandExt, FfmpegOut};
+use crate::process::{Chunks, CommandExt, FfmpegOut, cmd_err, exit_ok_stderr};
 use anyhow::Context;
 use log::{debug, info};
 use std::{path::Path, process::Stdio};
@@ -13,7 +13,7 @@ pub fn run(
     distorted: &Path,
     filter_complex: &str,
     fps: Option<f32>,
-) -> anyhow::Result<impl Stream<Item = VmafOut>> {
+) -> anyhow::Result<impl Stream<Item = VmafOut> + use<>> {
     info!(
         "vmaf {} vs reference {}",
         distorted.file_name().and_then(|n| n.to_str()).unwrap_or(""),
@@ -21,7 +21,8 @@ pub fn run(
     );
 
     let mut cmd = Command::new("ffmpeg");
-    cmd.arg2_opt("-r", fps)
+    cmd.kill_on_drop(true)
+        .arg2_opt("-r", fps)
         .arg2("-i", distorted)
         .arg2_opt("-r", fps)
         .arg2("-i", reference)
@@ -37,7 +38,9 @@ pub fn run(
 
     let cmd_str = cmd.to_cmd_str();
     debug!("cmd `{cmd_str}`");
-    let mut vmaf = ProcessChunkStream::try_from(cmd).context("ffmpeg vmaf")?;
+    let mut vmaf = crate::process::child::AddOnDropChunkStream::from(
+        ProcessChunkStream::try_from(cmd).context("ffmpeg vmaf")?,
+    );
 
     Ok(async_stream::stream! {
         let mut chunks = Chunks::default();